requirements.txt
kafka-python>=2.0.2
pyspark>=3.3.0
psycopg2-binary>=2.9.6
pandas>=2.0.0
pyarrow>=12.0.0
apache-airflow==2.6.3
faker>=18.9.0
pytest>=7.3.0
